Move UserMailer.deduplicated_send_match_notify to a MatchMailerService
