+++ app/controllers/matches_controller.rb
@@ -6,7 +6,7 @@ class MatchesController < ApplicationController
     authorize @match
     previous_status = @match.status
     @match.update status: params[:status]
-    UserMailer.deduplicated_send_match_notify(@match, previous_status)
+    MatchMailerService.deduplicated_notify_status(@match, previous_status)
     if @match.status_taking_care?
       CompanyMailer.notify_taking_care(@match).deliver_later
     end
